refactor(utils): remove language filtering and move files into `_hooks` folder
